+++ packages/cli/test/unit/commands/help.test.ts
@@ -470,6 +470,26 @@ describe('help command', () => {
         ).toMatchSnapshot();
       });
     });
+    describe('connex attach subcommand', () => {
+      it('connex attach subcommand help column width 120', () => {
+        expect(
+          help(connex.attachSubcommand, {
+            columns: 120,
+            parent: connex.connexCommand,
+          })
+        ).toMatchSnapshot();
+      });
+    });
+    describe('connex detach subcommand', () => {
+      it('connex detach subcommand help column width 120', () => {
+        expect(
+          help(connex.detachSubcommand, {
+            columns: 120,
+            parent: connex.connexCommand,
+          })
+        ).toMatchSnapshot();
+      });
+    });
   });
 
   describe('integration help output snapshots', () => {