+++ packages/python/src/index.ts
@@ -22,6 +22,7 @@ import {
   Span,
   BUILDER_INSTALLER_STEP,
   BUILDER_COMPILE_STEP,
+  BUILDER_PRE_DEPLOY_STEP,
   type BuildOptions,
   type GlobOptions,
   type BuildVX,
@@ -70,9 +71,12 @@ const PYTHON_ENTRYPOINT_DOCS_URL =
 
 import {
   detectPythonEntrypoint,
+  entrypointToModule,
   type DetectedPythonEntrypoint,
   type PythonEntrypoint,
 } from './entrypoint';
+
+export { detectEntrypoint } from './entrypoint';
 
 export const version = -1;
 
@@ -255,6 +259,7 @@ export const build: BuildVX = async ({
   config,
   span: parentSpan,
   service,
+  registerPreDeploy,
 }) => {
   let entrypoint: string | undefined =
     rawEntrypoint === '<detect>' ? undefined : rawEntrypoint;
@@ -630,7 +635,7 @@ export const build: BuildVX = async ({
   await uv.pip({
     venvPath,
     projectDir: join(workPath, entryDirectory),
-    args: ['install', ...pipPlatformArgs, runtimeDep],
+    args: ['install', '--link-mode', 'copy', ...pipPlatformArgs, runtimeDep],
   });
 
   if (shouldInstallVercelWorkers) {
@@ -642,7 +647,7 @@ export const build: BuildVX = async ({
     await uv.pip({
       venvPath,
       projectDir: join(workPath, entryDirectory),
-      args: ['install', ...pipPlatformArgs, workersDep],
+      args: ['install', '--link-mode', 'copy', ...pipPlatformArgs, workersDep],
     });
   }
 
@@ -654,8 +659,30 @@ export const build: BuildVX = async ({
   if (quirksResult.buildEnv) {
     Object.assign(pythonEnv, quirksResult.buildEnv);
   }
+
+  // Register a pre-deploy command that will be fired in the end of the
+  // build process (if all builders including this one succeed)
+  const preDeployCommand = config?.preDeployCommand;
+  if (registerPreDeploy && typeof preDeployCommand === 'string') {
+    const capturedEnv = { ...pythonEnv };
+    const capturedCwd = workPath;
+    registerPreDeploy(async () => {
+      await builderSpan
+        .child(BUILDER_PRE_DEPLOY_STEP, {
+          preDeployCommand,
+        })
+        .trace(async () => {
+          console.log(`Running pre-deploy command: \`${preDeployCommand}\``);
+          await execCommand(preDeployCommand, {
+            env: capturedEnv,
+            cwd: capturedCwd,
+          });
+        });
+    });
+  }
+
   debug('Entrypoint is', entrypoint);
-  const moduleName = entrypoint.replace(/\//g, '.').replace(/\.py$/i, '');
+  const moduleName = entrypointToModule(entrypoint);
 
   if (handlerFunction) {
     const entrypointPath = join(workPath, entrypoint);