+++ packages/fastify/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @vercel/fastify
 
+## 0.1.81
+
+### Patch Changes
+
+- @vercel/node@5.8.2
+
+## 0.1.80
+
+### Patch Changes
+
+- @vercel/node@5.8.1
+
+## 0.1.79
+
+### Patch Changes
+
+- Updated dependencies [22f77b9]
+  - @vercel/node@5.8.0
+
 ## 0.1.78
 
 ### Patch Changes