+++ packages/h3/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @vercel/h3
 
+## 0.1.87
+
+### Patch Changes
+
+- @vercel/node@5.8.2
+
+## 0.1.86
+
+### Patch Changes
+
+- @vercel/node@5.8.1
+
+## 0.1.85
+
+### Patch Changes
+
+- Updated dependencies [22f77b9]
+  - @vercel/node@5.8.0
+
 ## 0.1.84
 
 ### Patch Changes