+++ packages/koa/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @vercel/koa
 
+## 0.1.61
+
+### Patch Changes
+
+- @vercel/node@5.8.2
+
+## 0.1.60
+
+### Patch Changes
+
+- @vercel/node@5.8.1
+
+## 0.1.59
+
+### Patch Changes
+
+- Updated dependencies [22f77b9]
+  - @vercel/node@5.8.0
+
 ## 0.1.58
 
 ### Patch Changes